+++ app/src/main/java/com/chicohan/currencyexchange/data/model/ExchangeRateResponse.kt
@@ -1,8 +1,9 @@
 package com.chicohan.currencyexchange.data.model
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 data class ExchangeRateResponse(
     @SerializedName("success") val success: Boolean,
-    @SerializedName("quotes") val rates: Map<String, Double>?
+    @SerializedName("quotes") val rates: Map<String, BigDecimal>?
 )
